Solucionando warnings de Damn

+++ src/Damn/EnemyAttack.cpp
@@ -32,9 +32,9 @@ void damn::EnemyAttack::Update(float t)
 	if (_player && !_player->GetComponent<PlayerHealth>()->isDead()) {
 		eden_utils::Vector3 newPlayePos(_playerTr->GetPosition().GetX(), _tr->GetPosition().GetY(), _playerTr->GetPosition().GetZ());
 		eden_utils::Vector3 dir = (newPlayePos - _tr->GetPosition()).Normalized();
-		float frontAngle = acos(dir.Dot(_tr->GetForward())) * (180/PI); //Calcula el ángulo que forma el jugador con el forward del enemigo 
+		float frontAngle = acos((float)(dir.Dot(_tr->GetForward()))) * (180.0f/(float)(PI)); //Calcula el ángulo que forma el jugador con el forward del enemigo 
 		if (frontAngle > 1) { //Si el jugador está a un a más de un grado de diferencia, el enemigo sigue rotando
-			float rightAngle = acos(dir.Dot(_tr->GetRight())) * (180 / PI); //Usando el vector derecha del enemigo determinamos el lado hacia el que tiene que rotar
+			float rightAngle = acos((float)(dir.Dot(_tr->GetRight()))) * (180.0f / (float)(PI)); //Usando el vector derecha del enemigo determinamos el lado hacia el que tiene que rotar
 			if (rightAngle > 90) _rotationCoef = -1;
 			else _rotationCoef = 1;
 			_tr->Yaw(_rotationSpeed * _rotationCoef * t);
